# net: halve the allocs in ParseCIDR by sharing slice backing