@cstockton 2019-11-03
I've thought about this a good bit since I've opened it and have decided to withdraw this pull request. I created a library at the time of this pull request for internal use that could ParseCIDR into an existing buffer and fetched [CIDR masks](https://gist.github.com/cstockton/02638ab68bb06f4171da402b3bc8f56f) from a lookup table. The fact I wouldn't stop using it in favor of the stdlib for a pull req I made due to the memory cost just isn't sitting right with me, and I don't think any variation (1, 2, or 3 allocs) will under the constraints of the existing API.

If the Go team would consider an alternative pull request that added some public methods on those types I would be happy to implement an adaption of my patches / internal lib, i.e.:

```Go
// Init is like CIDRMask, but puts the mask into m(.| and returns the possibly extended m.)
func (m *IPMask) Init(ones, bits int)
func (m IPMask) Init(ones, bits int) IPMask

// will reuse the existing IP and IPMask, but removes the ability to get designated host
func (n *IPNet) (ParseCIDR|Parse|InitCIDR|???)(cidr string) error

// you could supplement that with something like the below, but feels a little off.
func (ip *IP) ParseCIDR() error
```

None of my existing thoughts about how to bolt onto the existing API are very elegant. Maybe something else will come to me and I'll open a separate proposal for it, or someone else has a better idea and is free to tag me with it.